Move to servlets.flight package.
