+++ flightmate/src/main/java/com/flightmate/servlets/flight/FlightStatisticsServlet.java
@@ -1,4 +1,4 @@
-package com.flightmate.servlets;
+package com.flightmate.servlets.flight;
 
 import com.flightmate.dao.FlightDao;
 import com.google.gson.Gson;
